requirements.txt
click==8.0.3
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
protobuf==3.19.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
zipp==3.6.0
